:recycle: refactor(chamaos_em_um_dia): separar dashboard em arquivos
